Change total items from int to long

+++ Shared/Common/General/PaginatedList.cs
@@ -5,11 +5,11 @@
         public IEnumerable<T> Items { get; set; }
         public int PageIndex { get; set; }
         public int TotalPages { get; set; }
-        public int TotalItems { get; set; }
+        public long TotalItems { get; set; }
         public bool HasPrevious => PageIndex > 0;
         public bool HasNext => PageIndex < TotalPages - 1;
 
-        public PaginatedList(IEnumerable<T> items, int count, int pageIndex, int pageSize)
+        public PaginatedList(IEnumerable<T> items, long count, int pageIndex, int pageSize)
         {
             PageIndex = pageIndex;
             TotalItems = count;
@@ -17,7 +17,7 @@
             Items = items;
         }
 
-        public static PaginatedList<T> Create(IEnumerable<T> items, int count, int pageIndex, int pageSize)
+        public static PaginatedList<T> Create(IEnumerable<T> items, long count, int pageIndex, int pageSize)
         {
             return new PaginatedList<T>(items, count, pageIndex, pageSize);
         }
